Segregated parsers-related files
