+++ spartsi/src/parser/impl/node/node_end.hpp
@@ -1,19 +1,19 @@
 #include "langspec/langspec.hpp"
 #include "builder/tree-builder.hpp"
 
-#include "util.hpp"
+#include "parser/impl/util.hpp"
 
 namespace spartsi {
     namespace parser {
         template<typename Str>
-        bool node_begin(builder::tree_builder &bld, langspec<Str> spec, const Str &line) {
+        bool node_end(builder::tree_builder &bld, langspec<Str> spec, const Str &line) {
             using namespace util;
             std::smatch m;
 
-            std::string pattern = spec.node.begin + " " + capture_quoted;
+            std::string pattern = spec.node.end;
 
             if(std::regex_match(line, m, weak_reg(pattern))) {
-                bld.node(m[1]);
+                bld.end_node();
                 return true;
             }
