+++ spartsi/test/parser/impl/node/node_begin.cpp
@@ -1,6 +1,6 @@
 #include "catch/catch.hpp"
 
-#include "parser/impl/node_begin.hpp"
+#include "parser/impl/node/node_begin.hpp"
 #include "langspec/default.hpp"
 using namespace std;
